+++ python/qx_utilities/processing/core.py
@@ -196,6 +196,22 @@ def use_or_skip_bold(sinfo, options, r=""):
     return bolds, bskip, len(bskip), r
 
 
+def get_bold_names(boldinfo, options):
+    """
+    Get bold names based on the boldinfo and options.
+    """
+
+    if "filename" in boldinfo and options["hcp_filename"] == "userdefined":
+        printbold = boldinfo["filename"]
+        boldtarget = boldinfo["filename"]
+        boldsource = boldinfo["filename"]
+    else:
+        printbold = str(boldinfo["bold_number"])
+        boldsource = "BOLD_%d" % (boldinfo["bold_number"])
+        boldtarget = "%s%s" % (options["hcp_bold_prefix"], printbold)
+
+    return printbold, boldtarget, boldsource
+
 def doOptionsCheck(options, sinfo, command):
     # logs
     logs = [e.strip() for e in re.split(r" +|\||, *", options["log"])]